+++ fundu/src/custom/builder.rs
@@ -255,6 +255,31 @@ impl<'a> CustomDurationParserBuilder<'a> {
         self
     }
 
+    /// Allow one or more delimiters between the leading sign and the number.
+    ///
+    /// See also [`crate::DurationParser::allow_sign_delimiter`].
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// use fundu::TimeUnit::*;
+    /// use fundu::{CustomDurationParserBuilder, CustomTimeUnit, Duration};
+    ///
+    /// let parser = CustomDurationParserBuilder::new()
+    ///     .time_units(&[CustomTimeUnit::with_default(NanoSecond, &["ns"])])
+    ///     .allow_sign_delimiter(|byte| byte.is_ascii_whitespace())
+    ///     .build();
+    ///
+    /// assert_eq!(parser.parse("+123ns"), Ok(Duration::positive(0, 123)));
+    /// assert_eq!(parser.parse("+\t\n 123ns"), Ok(Duration::positive(0, 123)));
+    /// assert_eq!(parser.parse("+ 123ns"), Ok(Duration::positive(0, 123)));
+    /// assert_eq!(parser.parse("+     123ns"), Ok(Duration::positive(0, 123)));
+    /// ```
+    pub const fn allow_sign_delimiter(mut self, delimiter: Delimiter) -> Self {
+        self.config.sign_delimiter = Some(delimiter);
+        self
+    }
+
     /// If set, parsing negative durations is possible
     ///
     /// See also [`crate::DurationParser::allow_negative`]
@@ -585,6 +610,12 @@ mod tests {
     fn test_custom_duration_parser_builder_when_allow_delimiter() {
         let builder = CustomDurationParserBuilder::new().allow_delimiter(|byte| byte == b' ');
         assert!(builder.config.allow_delimiter.unwrap()(b' '));
+    }
+
+    #[test]
+    fn test_custom_duration_parser_builder_when_allow_sign_delimiter() {
+        let builder = CustomDurationParserBuilder::new().allow_sign_delimiter(|byte| byte == b' ');
+        assert!(builder.config.sign_delimiter.unwrap()(b' '));
     }
 
     #[test]