+++ fundu/src/standard/parser.rs
@@ -287,6 +287,30 @@ impl<'a> DurationParser<'a> {
         self
     }
 
+    /// Allow one or more delimiters between the leading sign and the number.
+    ///
+    /// A [`Delimiter`] is defined as closure taking a byte and returning true if the delimiter
+    /// matched. If set to `None` (the default) no delimiter is allowed between the sign and the
+    /// number.
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// use fundu::{Duration, DurationParser};
+    ///
+    /// let mut parser = DurationParser::new();
+    /// parser.allow_sign_delimiter(Some(|byte| byte.is_ascii_whitespace()));
+    ///
+    /// assert_eq!(parser.parse("+123ns"), Ok(Duration::positive(0, 123)));
+    /// assert_eq!(parser.parse("+\t\n 123ns"), Ok(Duration::positive(0, 123)));
+    /// assert_eq!(parser.parse("+ 123ns"), Ok(Duration::positive(0, 123)));
+    /// assert_eq!(parser.parse("+     123ns"), Ok(Duration::positive(0, 123)));
+    /// ```
+    pub fn allow_sign_delimiter(&mut self, delimiter: Option<Delimiter>) -> &mut Self {
+        self.inner.config.sign_delimiter = delimiter;
+        self
+    }
+
     /// If true, then parsing negative durations is possible
     ///
     /// Without setting this option the parser returns [`ParseError::NegativeNumber`] if it
@@ -595,6 +619,13 @@ mod tests {
         let mut parser = DurationParser::new();
         parser.allow_delimiter(Some(|byte| byte == b' '));
         assert!(parser.inner.config.allow_delimiter.unwrap()(b' '));
+    }
+
+    #[test]
+    fn test_duration_parser_setting_allow_sign_delimiter() {
+        let mut parser = DurationParser::new();
+        parser.allow_sign_delimiter(Some(|byte| byte == b' '));
+        assert!(parser.inner.config.sign_delimiter.unwrap()(b' '));
     }
 
     #[test]